misc/requirements.txt
pillow
cairosvg
requests